drafting refactor of lithology module
