+++ quick_pp/lithology/carbonate.py
@@ -1,0 +1,2 @@
+class Carbonate:
+    pass
